api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.12",
    "langchain-core>=0.3.59",
    "langchain-huggingface>=0.2.0",
    "langchain-mongodb>=0.6.1",
    "langchain-openai>=0.3.16",
    "langgraph>=0.4.3",
    "langgraph-checkpoint-mongodb>=0.1.3",
    "loguru>=0.7.3",
    "opik>=1.7.18",
    "sentence-transformers>=4.1.0",
]
